@@ -1,7 +1,4 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
-
-import 'package:flutter/services.dart';
 import 'package:voice_changer_plugin/voice_changer_plugin.dart';
 
 void main() => runApp(MyApp());
@@ -12,9 +9,7 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
-
   // Platform messages are asynchronous, so we initialize in an async method.
-
 
   @override
   Widget build(BuildContext context) {
@@ -24,7 +19,12 @@ class _MyAppState extends State<MyApp> {
           title: const Text('Plugin example app'),
         ),
         body: Center(
-          child:FlatButton(onPressed: ()=>VoiceChangerPlugin.change("/Users/kayshen/Library/Developer/CoreSimulator/Devices/FBE19BC2-E192-4067-A370-30B699D1813D/data/Containers/Bundle/Application/6E9FBF91-41B8-47E4-B64F-FFD8BF284848/Runner.app/singing.wav", 1), child: Text("测试"))
+          child: ElevatedButton(
+            onPressed: () => VoiceChangerPlugin.change(
+                "/Users/kayshen/Library/Developer/CoreSimulator/Devices/FBE19BC2-E192-4067-A370-30B699D1813D/data/Containers/Bundle/Application/6E9FBF91-41B8-47E4-B64F-FFD8BF284848/Runner.app/singing.wav",
+                1),
+            child: Text("测试"),
+          ),
         ),
       ),
     );
